Add endpoint for reading pipette model info and organize endpoint files
