+++ .gitignore
@@ -2,6 +2,7 @@
 __pycache__/
 *.py[cod]
 *$py.class
+*.pytest_cache
 
 # C extensions
 *.so
